add polygon image modifiers

+++ Source/TWSPainter/PaintPointer.cs
@@ -11,7 +11,7 @@ namespace ASP
         static private string _pointerModel = "ASP/TheWriteStuff/Assets/paint-brush";
         static private float _offset = 0.2f;
         static private Vector3 _rotation = new Vector3(0f, -1f, 0f);
-        static float _maxDist = 5.0f;
+        static double _maxDist = 5.0f;
 
         private ASPPainter _painter;
         private GameObject _pointer = null;
